Create pet.lua
This was one of the last projects I've made for ComputerCraft. It uses co-routines to manage pet atrophy, animation, and user input. It's one of the more comprehensive programs I've made for CC v1.3.2. Pet files are saved to the computer to be loaded at a later time and are encrypted so the user cannot cheat. It's like a tamagotchi but for CC.